// packages/root/projects/root/src/lib/model/avatar.type.ts
export type AvatarModifier =
  | 'red'
  | 'orange'
  | 'amber'
  | 'yellow'
  | 'lime'
  | 'green'
  | 'emerald'
  | 'teal'
  | 'cyan'
  | 'sky'
  | 'blue'
  | 'indigo'
  | 'violet'
  | 'purple'
  | 'fuchsia'
  | 'pink'
  | 'rose'
  | 'slate'
  | 'gray'
  | 'zinc'
  | 'neutral'
  | 'stone';

export type AvatarSize = 'small' | 'medium' | 'large';

export type AvatarShape = 'square' | 'rounded' | 'circle';